# Enable fastpath on invokesuper